Initialize the extra values to zero
When loading an ASCII STL file, the facet extra values are set to random values.

This is because the facet is an uninitialized structure on the stack.
During the load process of a binary STL, these values will be read correctly.
But these values are skipped during the load process of an ASCII STL,
leaving the uninitialized values untouched.

Situations where an ASCII STL is read, but a Binary STL is written would then
result in different output across multiple runs.

An easy fix is to simply set the extra values to zero at the beginning of the
load process. If the load is binary, that value will be overwritten for each
facet. If the load is ASCII then the zero value can be correctly used for each facet.

Fixes https://github.com/admesh/admesh/issues/32
Merges https://github.com/admesh/admesh/pull/31